requirements.txt
beautifulsoup4 
request 
spotipy 
lyricsgenius 
python-musicbrainzngs 
cohere
dotenv
services
typing 
utils
chromadb
flask